Add GetUserInfo endpoint and user fields
- Added new properties (FirstName, LastName, Gender, DateOfBirth) to the User entity.
- Created migration to update the database schema.
- Updated AuthenticationService to handle new user fields during registration.
- Implemented GetUserInfo endpoint in AuthController with IUserInfoService.
- Created UserInfoService and UserInfoResDto for structured user responses.
- Configured JWT authentication and updated DI for new services.
- Refactored GenericRepository to adjust GetByIdAsync method.

+++ ReelsCommerceSystem.Application/Interfaces/Repositories/IGenericRepository.cs
@@ -9,7 +9,7 @@ public interface IGenericRepository<T> where T : BaseEntity
     public Task<IReadOnlyList<T>> GetAllAsync();
     public Task<IReadOnlyList<T>> GetAllWithSpecAsync(ISpecification<T> specification);
 
-    public Task<T?> GetByIdAsync(long id);
+    public Task<T?> GetByIdAsync(int id);
     public Task<T?> GetWithSpecAsync(ISpecification<T> specification);
 
     public Task AddAsync(T entity);
